# design/API.yaml
openapi: 3.0.0
info:
  title: ViSOL
  description: >-
    Visual Interactive Optimization Tool. You can view this documentation over
    at [Apiary](https://visol.docs.apiary.io).
  contact: { }
  version: 1.0.0
servers:
  - url: 'http://localhost:8080/visol/rest'
    variables: { }
paths:
...
  '/vessels/{vessel_id}':
    parameters:
      - name: vessel_id
        in: path
        description: ID of the Vessel.
        required: true
        schema:
          $ref: '#/components/schemas/VesselID'
    get:
      tags:
        - Vessels
      summary: Retrieve a Vessel
      operationId: retrieveVessel
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vessel'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Vessels
      summary: Replace a Vessel
      operationId: replaceVessel
      requestBody:
        description: ''
        headers:
          Reason:
            schema:
              $ref: '#/components/schemas/Reason'
            description: The reason for which the Vessel was replaced/updated.
            example: Additional cargo (#2524)
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vessel'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vessel'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Vessels
      summary: Delete a Vessel
      operationId: deleteVessel
      responses:
        '204':
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
...
  /terminals:
    post:
      tags:
        - Terminals
      summary: Create a Terminal
      operationId: createTerminal
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Terminal'
        required: true
      responses:
        '201':
          description: ''
          headers:
            Location:
              schema:
                type: string
                pattern: '^\/terminals\/[0-9]{1,10}$'
                maxLength: 21
                example: /terminals/1
              description: The URL of the newly created Terminal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Terminal'
  '/terminals/{terminal_id}':
    parameters:
      - name: terminal_id
        in: path
        description: ID of the Terminal.
        required: true
        schema:
          $ref: '#/components/schemas/TerminalID'
    get:
      tags:
        - Terminals
      summary: Retrieve a Terminal
      operationId: retrieveTerminal
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Terminal'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Terminals
      summary: Replace a Terminal
      operationId: replaceTerminal
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Terminal'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Terminal'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Terminals
      summary: Delete a Terminal
      operationId: deleteTerminal
      responses:
        '204':
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
...
  '/changes/vessels/{vessel_id}':
    parameters:
      - name: vessel_id
        in: path
        description: ID of the Vessel.
        required: true
        schema:
          $ref: '#/components/schemas/VesselID'
    get:
      tags:
        - Changes
      summary: Retrieve a Vessel's Changes
      operationId: retrieveVesselChanges
      parameters:
        - name: from
          in: query
          description: Time to start looking from
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: to
          in: query
          description: Time to end looking at
          schema:
            $ref: '#/components/schemas/Timestamp'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VesselChanges'
    delete:
      tags:
        - Changes
      summary: Delete a Vessel's Changes
      operationId: deleteVesselChanges
      parameters:
        - name: from
          in: query
          description: Time to start deleting from
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: to
          in: query
          description: Time to end deleting at
          schema:
            $ref: '#/components/schemas/Timestamp'
      responses:
        '204':
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
  /changes/vessels:
    get:
      tags:
        - Changes
      summary: Retrieve all Vessels' Changes
      operationId: retrieveAllVesselChanges
      parameters:
        - name: from
          in: query
          description: Time to start looking from
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: to
          in: query
          description: Time to end looking at
          schema:
            $ref: '#/components/schemas/Timestamp'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VesselChangesMap'
  '/changes/vessels/{vessel_id}/{date}':
    parameters:
      - name: vessel_id
        in: path
        description: ID of the Vessel.
        required: true
        schema:
          $ref: '#/components/schemas/VesselID'
      - name: date
        in: path
        description: >-
          Time at which the change was made. Can be formatted without colons and
          dashes.
        required: true
        schema:
          $ref: '#/components/schemas/Timestamp'
    get:
      tags:
        - Changes
      summary: Retrieve a Vessel's Change
      operationId: retrieveVesselChange
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VesselChange'
    delete:
      tags:
        - Changes
      summary: Delete a Vessel's Change
      operationId: deleteVesselChange
      responses:
        '204':
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
  '/changes/schedules/{vessel_id}':
    parameters:
      - name: vessel_id
        in: path
        description: ID of the Vessel whose Schedule should be changed.
        required: true
        schema:
          $ref: '#/components/schemas/VesselID'
    get:
      tags:
        - Changes
      summary: Retrieve all Schedule's Changes
      operationId: retrieveAllScheduleChanges
      parameters:
        - name: from
          in: query
          description: Time to start looking from
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: to
          in: query
          description: Time to end looking at
          schema:
            $ref: '#/components/schemas/Timestamp'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleChanges'
    delete:
      tags:
        - Changes
      summary: Delete a Schedule's Changes
      operationId: deleteScheduleChanges
      parameters:
        - name: from
          in: query
          description: Time to start deleting from
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: to
          in: query
          description: Time to end deleting at
          schema:
            $ref: '#/components/schemas/Timestamp'
      responses:
        '204':
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
  /changes/schedules:
    get:
      tags:
        - Changes
      summary: Retrieve all Schedules' Changes
      operationId: retrieveAllSchedulesChanges
      parameters:
        - name: from
          in: query
          description: Time to start looking from
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: to
          in: query
          description: Time to end looking at
          schema:
            $ref: '#/components/schemas/Timestamp'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleChangesMap'
  '/changes/schedules/{vessel_id}/{date}':
    parameters:
      - name: vessel_id
        in: path
        description: ID of the Vessel which Schedule should be changed.
        required: true
        schema:
          $ref: '#/components/schemas/VesselID'
      - name: date
        in: path
        description: >-
          Time at which the change was made. Can be formatted without colons and
          dashes.
        required: true
        schema:
          $ref: '#/components/schemas/Timestamp'
    get:
      tags:
        - Changes
      summary: Retrieve a Schedule's Change
      operationId: retrieveScheduleChange
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleChange'
    delete:
      tags:
        - Changes
      summary: Delete a Schedule's Change
      operationId: deleteScheduleChange
      responses:
        '204':
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
security:
  - bearerAuth: [ ]
components:
  schemas:
    Reason:
      type: string
      pattern: '^(?:.|\s)*$'
      maxLength: 1073741824
      nullable: true
      example: New vessel inbound (#142)
...
    NullableTimestamp:
      allOf:
        - $ref: '#/components/schemas/Timestamp'
        - nullable: true
    Timestamp:
      type: string
      format: date-time
      pattern: >-
        ^(-?[0-9]+)-?(1[0-2]|0[1-9])-?(3[01]|0[1-9]|[12][0-9])T(2[0-3]|[01][0-9]):?([0-5][0-9]):?([0-5][0-9])Z$
      maxLength: 22
      example: 2022-05-11T16:32:35Z
    NullableSchedule:
      allOf:
        - $ref: '#/components/schemas/Schedule'
        - title: Nullable Schedule
          nullable: true
    Schedule:
      title: Schedule
      description: >-
        A Schedule object has the following attributes:


        + vessel: `1` (integer) - ID of the Vessel.
        
        + manual: `false` (`boolean`) - True if the Vessel should be planned at the supplied Berth and timestamp or false if it should be planned by the software.
        
        + berth: `20` (integer) - ID of the Berth the Vessel should be unloaded at.
        
        + start: `2022-05-12T07:23:00Z` (timestamp) - When the Vessel should start being unloaded.
        
        + end: `2022-05-13T21:41:00Z` (timestamp) - When the Vessel is expected to end being unloaded.
        
        
        The `berth` and `start` attributes are not necessary if the `type` is `automatic`.
        The `end` attribute will be calculated by the server.
      type: object
      properties:
        vessel:
          $ref: '#/components/schemas/VesselID'
        manual:
          type: boolean
        berth:
          $ref: '#/components/schemas/BerthID'
        start:
          $ref: '#/components/schemas/Timestamp'
        end:
          $ref: '#/components/schemas/Timestamp'
      additionalProperties: false
      required:
        - vessel
        - manual
        - berth
        - start
      example:
        vessel: 1
        manual: true
        berth: 30
        start: 2022-05-11T16:32:00Z
        end: 2022-05-11T21:26:00Z
...
    VesselID:
      title: Vessel ID
      description: >-
        An ID of a Vessel. Vessel IDs are unique, and are generated by the
        database.
      type: integer
      format: int32
      minimum: 1
      maximum: 2147483647
      example: 1
...
    NullableVessel:
      allOf:
        - $ref: '#/components/schemas/Vessel'
        - title: Nullable Vessel
          nullable: true
    Vessel:
      title: Vessel
      description: >-
        A Vessel object has the following attributes:


        + name: `BigBoatHan` (string) - The name of the Vessel.

        + arrival: `2022-05-11T13:10:00Z` (timestamp) - When the Vessel is expected to arrive.

        + deadline: `2022-05-12T09:54:00Z` (timestamp) - When the Vessel must be completely unloaded.

        + containers: `5` (integer) - The amount of containers on the Vessel to be unloaded.

        + cost_per_hour: `2.6` (decimal) - The cost per hour from arrival till fully unloaded.

        + destination: `3` (integer) - The terminal where the Vessel should be unloaded.

        + length: `30` (integer) - The length of the Vessel in metres.

        + width: `10` (integer) - The width of the Vessel in metres.

        + depth: `20` (integer) - The depth of the Vessel in metres.
      type: object
      properties:
        name:
          type: string
          pattern: '^(?:.|\s)+$'
          maxLength: 64
        arrival:
          $ref: '#/components/schemas/Timestamp'
        deadline:
          $ref: '#/components/schemas/NullableTimestamp'
        containers:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
        cost_per_hour:
          type: number
          format: real
          minimum: 0
          maximum: 1e+37
        destination:
          $ref: '#/components/schemas/TerminalID'
        length:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
        width:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
        depth:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
      additionalProperties: false
      required:
        - name
        - arrival
        - containers
        - cost_per_hour
        - destination
        - length
        - width
        - depth
      example:
        name: Vessel 1
        arrival: 2022-05-11T13:56:00Z
        deadline: 2022-05-12T18:13:00Z
        containers: 30
        cost_per_hour: 2.6
        destination: 4
        length: 30
        width: 10
        depth: 30
...
    BerthID:
      title: Berth ID
      description: >-
        An ID of a Berth. Berth IDs are unique, and are generated by the
        database.
      type: integer
      format: int32
      minimum: 1
      maximum: 2147483647
      example: 1
...
    TerminalID:
      title: Terminal ID
      description: >-
        An ID of a Terminal. Terminal IDs are unique, and are generated by the
        database.
      type: integer
      format: int32
      minimum: 1
      maximum: 2147483647
      example: 1
    Terminal:
      title: Terminal
      description: |-
        A Terminal object has the following attributes:
        
        
        + port: `2` (integer) - The Port where the Terminal is located.
        
        + name: `Han & Co. Ltd.` (string) - The name of the Terminal.
      type: object
      properties:
        port:
          $ref: '#/components/schemas/PortID'
        name:
          type: string
          pattern: '^(?:.|\s)+$'
          maxLength: 64
      additionalProperties: false
      required:
        - port
        - name
      example:
        port: 2
        name: Han & Co. Ltd.
...
    PortID:
      title: Port ID
      description: 'An ID of a Port. Port IDs are unique, and are generated by the database.'
      type: integer
      format: int32
      minimum: 1
      maximum: 2147483647
      example: 1
...
    VesselChange:
      title: Vessel Change
      description: >-
        A Vessel Change object has the following attributes:


        + vessel: `1` (integer) - The ID of the Vessel.

        + date: `2022-05-11T16:32:35Z` (timestamp) - The date at which the change was made.

        + old: `{"name":"Vessel 1","arrival":"2022-05-11T13:56:00.000Z","deadline":"2022-05-12T18:13:00.000Z","containers":30,"cost_per_hour":2.6,"destination":4,"length":30,"width":10,"depth":30}` (JSON) - The Vessel before being changed.

        + new: `{"name":"New Vessel 1","arrival":"2022-05-11T13:56:00.000Z","deadline":"2022-05-12T18:13:00.000Z","containers":40,"cost_per_hour":2.6,"destination":4,"length":30,"width":10,"depth":30}` (JSON) - The Vessel after being changed.

        + reason: `Additional cargo (#2524)` (string) - The reason for the change.
      type: object
      properties:
        vessel:
          $ref: '#/components/schemas/VesselID'
        date:
          $ref: '#/components/schemas/Timestamp'
        old:
          $ref: '#/components/schemas/NullableVessel'
        new:
          $ref: '#/components/schemas/Vessel'
        reason:
          $ref: '#/components/schemas/Reason'
      additionalProperties: false
      required:
        - vessel
        - date
        - old
        - new
        - reason
      example:
        vessel: 30
        date: 2022-05-11T16:32:35Z
        old:
          name: Vessel 1
          arrival: 2022-05-11T13:56:00Z
          deadline: 2022-05-12T18:13:00Z
          containers: 30
          cost_per_hour: 2.6
          destination: 4
          length: 30
          width: 10
          depth: 30
        new:
          name: New Vessel 1
          arrival: 2022-05-11T13:56:00Z
          deadline: 2022-05-12T18:13:00Z
          containers: 40
          cost_per_hour: 2.6
          destination: 4
          length: 30
          width: 10
          depth: 30
        reason: 'Additional cargo (#2524)'
    VesselChanges:
      title: Vessel Changes
      description: An array of Vessel Change objects.
      type: array
      items:
        $ref: '#/components/schemas/VesselChange'
      maxItems: 2147483647
      example:
        - vessel: 30
          date: 2022-05-11T16:32:35Z
          old:
            name: Vessel 1
            arrival: 2022-05-11T13:56:00Z
            deadline: 2022-05-12T18:13:00Z
            containers: 30
            cost_per_hour: 2.6
            destination: 4
            length: 30
            width: 10
            depth: 30
          new:
            name: New Vessel 1
            arrival: 2022-05-11T13:56:00Z
            deadline: 2022-05-12T18:13:00Z
            containers: 40
            cost_per_hour: 2.6
            destination: 4
            length: 30
            width: 10
            depth: 30
          reason: 'Additional cargo (#2524)'
        - vessel: 30
          date: 2022-05-11T21:26:24Z
          old:
            name: New Vessel 1
            arrival: 2022-05-11T13:56:00Z
            deadline: 2022-05-12T18:13:00Z
            containers: 40
            cost_per_hour: 2.6
            destination: 4
            length: 30
            width: 10
            depth: 30
          new:
            name: New Vessel 1
            arrival: 2022-05-18T13:56:00Z
            deadline: 2022-05-19T18:13:00Z
            containers: 40
            cost_per_hour: 2.6
            destination: 4
            length: 30
            width: 10
            depth: 30
          reason: 'Ship delayed (#3517)'
    VesselChangesMap:
      title: Vessel Changes Map
      description: A map of Vessel Changes keyed by Vessel ID.
      type: object
      additionalProperties:
        $ref: '#/components/schemas/VesselChanges'
      example:
        '30':
          - vessel: 30
            date: 2022-05-11T16:32:35Z
            old:
              name: Vessel 1
              arrival: 2022-05-11T13:56:00Z
              deadline: 2022-05-12T18:13:00Z
              containers: 30
              cost_per_hour: 2.6
              destination: 4
              length: 30
              width: 10
              depth: 30
            new:
              name: New Vessel 1
              arrival: 2022-05-11T13:56:00Z
              deadline: 2022-05-12T18:13:00Z
              containers: 40
              cost_per_hour: 2.6
              destination: 4
              length: 30
              width: 10
              depth: 30
            reason: 'Additional cargo (#2524)'
          - vessel: 30
            date: 2022-05-11T21:26:24Z
            old:
              name: New Vessel 1
              arrival: 2022-05-11T13:56:00Z
              deadline: 2022-05-12T18:13:00Z
              containers: 40
              cost_per_hour: 2.6
              destination: 4
              length: 30
              width: 10
              depth: 30
            new:
              name: New Vessel 1
              arrival: 2022-05-18T13:56:00Z
              deadline: 2022-05-19T18:13:00Z
              containers: 40
              cost_per_hour: 2.6
              destination: 4
              length: 30
              width: 10
              depth: 30
            reason: 'Ship delayed (#3517)'
    ScheduleChange:
      title: Schedule Change
      description: >-
        A Vessel Change object has the following attributes:


        + vessel: `1` (integer) - The ID of the Schedule's Vessel. 

        + date: `2022-05-11T16:32:35Z` (timestamp) - The date at which the change was made.

        + old: `{"vessel":30,"manual":true,"berth":20,"start":"2022-05-15T16:32:00.000Z","end":"2022-05-15T21:26:00.000Z"}` (JSON) - The Schedule before being changed.

        + new: `{"vessel":30,"manual":false,"berth":20,"start":"2022-05-15T16:32:00.000Z","end":"2022-05-15T21:26:00.000Z"}` (JSON) - The Schedule after being changed.

        + reason: `Turn on automatic scheduling` (string) - The reason for the change.
      type: object
      properties:
        vessel:
          $ref: '#/components/schemas/VesselID'
        date:
          $ref: '#/components/schemas/Timestamp'
        old:
          $ref: '#/components/schemas/NullableSchedule'
        new:
          $ref: '#/components/schemas/Schedule'
        reason:
          $ref: '#/components/schemas/Reason'
      additionalProperties: false
      required:
        - vessel
        - date
        - old
        - new
        - reason
      example:
        vessel: 30
        date: 2022-05-11T16:32:35Z
        old:
          vessel: 30
          manual: true
          berth: 20
          start: 2022-05-15T16:32:00Z
          end: 2022-05-15T21:26:00Z
        new:
          vessel: 30
          manual: false
          berth: 20
          start: 2022-05-15T16:32:00Z
          end: 2022-05-15T21:26:00Z
        reason: Turn on automatic scheduling
    ScheduleChanges:
      title: Schedule Changes
      description: An array of Schedule Change objects.
      type: array
      items:
        $ref: '#/components/schemas/ScheduleChange'
      maxItems: 2147483647
      example:
        - vessel: 30
          date: 2022-05-11T16:32:35Z
          old:
            vessel: 30
            manual: true
            berth: 20
            start: 2022-05-15T16:32:00Z
            end: 2022-05-15T21:26:00Z
          new:
            vessel: 30
            manual: false
            berth: 20
            start: 2022-05-15T16:32:00Z
            end: 2022-05-15T21:26:00Z
          reason: Turn on automatic scheduling
        - vessel: 30
          date: 2022-05-11T21:26:24Z
          old:
            vessel: 30
            manual: false
            berth: 20
            start: 2022-05-15T16:32:00Z
            end: 2022-05-15T21:26:00Z
          new:
            vessel: 30
            manual: false
            berth: 18
            start: 2022-05-09T10:24:00Z
            end: 2022-05-10T16:17:00Z
          reason: Automatic scheduling job
    ScheduleChangesMap:
      title: Schedule Changes Map
      description: A map of Schedule Changes keyed by a Schedule's Vessel ID.
      type: object
      additionalProperties:
        $ref: '#/components/schemas/ScheduleChanges'
      example:
        '30':
          - vessel: 30
            date: 2022-05-11T16:32:35Z
            old:
              vessel: 30
              manual: true
              berth: 20
              start: 2022-05-15T16:32:00Z
              end: 2022-05-15T21:26:00Z
            new:
              vessel: 30
              manual: false
              berth: 20
              start: 2022-05-15T16:32:00Z
              end: 2022-05-15T21:26:00Z
            reason: Turn on automatic scheduling
          - vessel: 30
            date: 2022-05-11T21:26:24Z
            old:
              vessel: 30
              manual: false
              berth: 20
              start: 2022-05-15T16:32:00Z
              end: 2022-05-15T21:26:00Z
            new:
              vessel: 30
              manual: false
              berth: 18
              start: 2022-05-09T10:24:00Z
              end: 2022-05-10T16:17:00Z
            reason: Automatic scheduling job
  responses:
    NotFound:
      description: 404 not found
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...